+++ back/src/main/java/com/studypot/back/applications/UserService.java
@@ -2,14 +2,7 @@ package com.studypot.back.applications;
 
 import com.studypot.back.domain.User;
 import com.studypot.back.domain.UserRepository;
-import com.studypot.back.dto.user.ProfileProjection;
-import com.studypot.back.dto.user.ProfileResponseDto;
-import com.studypot.back.dto.user.UpdateProfileRequestDto;
 import com.studypot.back.exceptions.ExistEmailException;
-import com.studypot.back.exceptions.UnregisteredEmailException;
-import com.studypot.back.exceptions.UserNotFoundException;
-import com.studypot.back.exceptions.WrongPasswordException;
-import java.util.Optional;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -42,31 +35,4 @@ public class UserService {
     return userRepository.save(user);
   }
 
-  public User authenticate(String email, String password) {
-
-    User user = userRepository.findByEmail(email).orElseThrow(() -> new UnregisteredEmailException(email));
-
-    if (!passwordEncoder.matches(password, user.getPassword())) {
-      throw new WrongPasswordException();
-    }
-
-    return user;
-  }
-
-  public ProfileProjection getProfile(String name) {
-    return userRepository.findByName(name).orElseThrow(UserNotFoundException::new);
-  }
-
-  public ProfileResponseDto updateProfile(Long userId, UpdateProfileRequestDto updateProfileRequestDto) {
-    User updateUser = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-    ProfileResponseDto profileResponseDto = updateUser.updateProfile(updateProfileRequestDto);
-    userRepository.save(updateUser);
-
-    return profileResponseDto;
-  }
-
-  public User checkRefreshToken(Long userId) {
-    Optional<User> user = userRepository.findById(userId);
-    return user.orElseThrow(() -> new UnregisteredEmailException("User Not Found"));
-  }
 }
