+++ back/src/main/java/com/studypot/back/interfaces/SessionController.java
@@ -1,6 +1,6 @@
 package com.studypot.back.interfaces;
 
-import com.studypot.back.applications.UserService;
+import com.studypot.back.applications.SessionService;
 import com.studypot.back.auth.UserId;
 import com.studypot.back.domain.User;
 import com.studypot.back.dto.session.SessionRequestDto;
@@ -8,6 +8,7 @@ import com.studypot.back.dto.session.SessionResponseDto;
 import com.studypot.back.utils.JwtUtil;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,17 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @Api("세션")
+@RequiredArgsConstructor
 public class SessionController {
 
   private final JwtUtil jwtUtil;
 
-  private final UserService userService;
-
-  public SessionController(UserService userService, JwtUtil jwtUtil) {
-
-    this.userService = userService;
-    this.jwtUtil = jwtUtil;
-  }
+  private final SessionService sessionService;
 
   @PostMapping("/login")
   @ApiOperation("로그인")
@@ -34,7 +30,7 @@ public class SessionController {
     String email = resource.getEmail();
     String password = resource.getPassword();
 
-    User user = userService.authenticate(email, password);
+    User user = sessionService.authenticate(email, password);
 
     String accessToken = jwtUtil.createAccessToken(user.getId(), user.getName());
     String refreshToken = jwtUtil.createRefreshToken(user.getId());
@@ -50,7 +46,7 @@ public class SessionController {
   @ApiOperation("토큰 리프레시")
   public SessionResponseDto refreshToken(@UserId Long userId) {
 
-    User user = userService.checkRefreshToken(userId);
+    User user = sessionService.checkRefreshToken(userId);
 
     String userName = user.getName();
     String accessToken = jwtUtil.createAccessToken(userId, userName);
